requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cloudinary==1.29.0
Django==3.1.12
django-bootstrap-v5==1.0.11
django-bootstrap5==21.3
django-crispy-forms==1.14.0
django-emoji-picker==0.0.6
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.9
